+++ 03_Implementation/EditandoMundos/src/main/java/org/itson/presentacion/FrmEditarPublicaciones.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.logging.Logger;
 import javax.swing.JFrame;
 import javax.swing.table.DefaultTableModel;
-import org.itson.controladores.ControladorAutor;
-import org.itson.dominio.Autor;
+import org.itson.controladores.ControladorPublicacion;
+import org.itson.dominio.Publicacion;
 import org.itson.dominio.Usuario;
 import org.itson.utils.ConfiguracionPaginado;
 import org.itson.utils.Dialogs;
@@ -15,41 +15,37 @@ import org.itson.utils.FormUtils;
  *
  * @author Toled
  */
-public class EditarAutoresForm extends JFrameActualizable {
+public class FrmEditarPublicaciones extends JFrameActualizable{
 
-    private static final Logger LOG = Logger.getLogger(EditarAutoresForm.class.getName());
-    private ConfiguracionPaginado configPaginado;
+    private static final Logger LOG = Logger.getLogger(FrmEditarPublicaciones.class.getName());
     private Usuario clienteLoggeado;
+    private ConfiguracionPaginado configPaginado;
     private final JFrame frmAnterior;
-    private EditarAutorForm editarAutorForm;
-    private List<Autor> autores;
+    private FrmEditarPublicacion editarPublicacionForm;
+    private List<Publicacion> publicaciones;
 
-    public EditarAutoresForm(JFrame frmAnterior, Usuario clienteLoggeado) {
+    public FrmEditarPublicaciones(JFrame frmAnterior, Usuario clienteLoggeado) {
         initComponents();
         this.configPaginado = new ConfiguracionPaginado(this.tblPublicaciones.getModel().getRowCount(), 0);
         this.clienteLoggeado = clienteLoggeado;
         this.frmAnterior = frmAnterior;
-        this.initFormsConectados();
-        cargarTablaAutores();
+        cargarTablaPublicaciones();
+        initFormsConectados();
     }
 
-    public final void cargarTablaAutores() {
+    public final void cargarTablaPublicaciones() {
 
-        autores = conseguirListaAutores();
+        publicaciones = this.conseguirListaPublicaciones();
 
         DefaultTableModel modeloTabla = (DefaultTableModel) this.tblPublicaciones.getModel();
         modeloTabla.setRowCount(0);
-        String nombreCompleto;
-        for (Autor autor : autores) {
-            nombreCompleto = autor.getNombres() + " "
-                    + autor.getApellidoPaterno()
-                    + " " + autor.getApellidoMaterno();
-
+        for (Publicacion publicacion : publicaciones) {
             Object[] fila = {
-                nombreCompleto,
-                autor.getEdad(),
-                autor.getNacionalidad()
-            };
+                publicacion.getTitulo(),
+                publicacion.getAutor().getNombres() + " " + publicacion.getAutor().getApellidoPaterno(),
+                publicacion.getNoPaginas(),
+                "$" + publicacion.getCostoProd(),
+                "$" + publicacion.getCostoVenta()};
 
             modeloTabla.addRow(fila);
         }
@@ -67,8 +63,8 @@ public class EditarAutoresForm extends JFrameActualizable {
         panelTablaCuentas = new javax.swing.JScrollPane();
         tblPublicaciones = new javax.swing.JTable();
         btnAdelante = new javax.swing.JButton();
-        btnEditar = new javax.swing.JButton();
         btnRetroceder = new javax.swing.JButton();
+        btnEditar = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setMinimumSize(new java.awt.Dimension(600, 400));
@@ -82,7 +78,7 @@ public class EditarAutoresForm extends JFrameActualizable {
 
         lblOperaciones.setFont(new java.awt.Font("Nirmala UI Semilight", 0, 24)); // NOI18N
         lblOperaciones.setForeground(new java.awt.Color(255, 255, 255));
-        lblOperaciones.setText("Editar Autor");
+        lblOperaciones.setText("Editar Publicaciones");
 
         btnAtras.setBackground(new java.awt.Color(0, 102, 255));
         btnAtras.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
@@ -103,9 +99,9 @@ public class EditarAutoresForm extends JFrameActualizable {
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
                 .addGap(20, 20, 20)
                 .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(155, 155, 155)
+                .addGap(139, 139, 139)
                 .addComponent(lblOperaciones)
-                .addContainerGap(238, Short.MAX_VALUE))
+                .addContainerGap(174, Short.MAX_VALUE))
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -121,26 +117,26 @@ public class EditarAutoresForm extends JFrameActualizable {
 
         tblPublicaciones.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null}
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null}
             },
             new String [] {
-                "Nombre", "Edad", "Nacionalidad"
+                "Titulo", "Autor", "No. Paginas", "Producción", "Venta"
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.String.class, java.lang.String.class, java.lang.String.class
+                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class
             };
             boolean[] canEdit = new boolean [] {
-                false, false, false
+                false, false, false, false, true
             };
 
             public Class getColumnClass(int columnIndex) {
@@ -153,11 +149,6 @@ public class EditarAutoresForm extends JFrameActualizable {
         });
         panelTablaCuentas.setViewportView(tblPublicaciones);
         tblPublicaciones.getColumnModel().getSelectionModel().setSelectionMode(javax.swing.ListSelectionModel.SINGLE_INTERVAL_SELECTION);
-        if (tblPublicaciones.getColumnModel().getColumnCount() > 0) {
-            tblPublicaciones.getColumnModel().getColumn(0).setResizable(false);
-            tblPublicaciones.getColumnModel().getColumn(1).setResizable(false);
-            tblPublicaciones.getColumnModel().getColumn(2).setResizable(false);
-        }
 
         Background.add(panelTablaCuentas, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 150, 560, 190));
 
@@ -174,19 +165,6 @@ public class EditarAutoresForm extends JFrameActualizable {
         });
         Background.add(btnAdelante, new org.netbeans.lib.awtextra.AbsoluteConstraints(540, 340, 30, 30));
 
-        btnEditar.setBackground(new java.awt.Color(0, 102, 255));
-        btnEditar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
-        btnEditar.setForeground(new java.awt.Color(255, 255, 255));
-        btnEditar.setText("Editar");
-        btnEditar.setBorder(null);
-        btnEditar.setBorderPainted(false);
-        btnEditar.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnEditarActionPerformed(evt);
-            }
-        });
-        Background.add(btnEditar, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 360, 140, 30));
-
         btnRetroceder.setBackground(new java.awt.Color(0, 102, 255));
         btnRetroceder.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
         btnRetroceder.setForeground(new java.awt.Color(255, 255, 255));
@@ -199,6 +177,19 @@ public class EditarAutoresForm extends JFrameActualizable {
             }
         });
         Background.add(btnRetroceder, new org.netbeans.lib.awtextra.AbsoluteConstraints(510, 340, 30, 30));
+
+        btnEditar.setBackground(new java.awt.Color(0, 102, 255));
+        btnEditar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
+        btnEditar.setForeground(new java.awt.Color(255, 255, 255));
+        btnEditar.setText("Editar");
+        btnEditar.setBorder(null);
+        btnEditar.setBorderPainted(false);
+        btnEditar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnEditarActionPerformed(evt);
+            }
+        });
+        Background.add(btnEditar, new org.netbeans.lib.awtextra.AbsoluteConstraints(250, 350, 150, 30));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -229,22 +220,21 @@ public class EditarAutoresForm extends JFrameActualizable {
      */
     private void btnAdelanteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAdelanteActionPerformed
         this.configPaginado.avanzarPag();
-        this.cargarTablaAutores();
+        this.cargarTablaPublicaciones();
     }//GEN-LAST:event_btnAdelanteActionPerformed
     /**
      * Retrocede en la pagina de operaciones
      *
      * @param evt Evento que lo acciono
      */
-    private void btnEditarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEditarActionPerformed
-        this.editarElementoSeleccionado();
-        this.cargarTablaAutores();
-    }//GEN-LAST:event_btnEditarActionPerformed
-
     private void btnRetrocederActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRetrocederActionPerformed
         this.configPaginado.retrocederPag();
-        this.cargarTablaAutores();
+        this.cargarTablaPublicaciones();
     }//GEN-LAST:event_btnRetrocederActionPerformed
+
+    private void btnEditarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEditarActionPerformed
+        this.editarElementoSeleccionado();
+    }//GEN-LAST:event_btnEditarActionPerformed
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
@@ -259,31 +249,31 @@ public class EditarAutoresForm extends JFrameActualizable {
     private javax.swing.JTable tblPublicaciones;
     // End of variables declaration//GEN-END:variables
 
-    private List<Autor> conseguirListaAutores() {
-        return ControladorAutor.consultaPaginado(this.configPaginado);
+    private List<Publicacion> conseguirListaPublicaciones() {
+        return ControladorPublicacion.consultaPaginado(this.configPaginado);
     }
 
     private void regresar() {
         FormUtils.regresar(frmAnterior, this);
     }
-
+    
     private void editarElementoSeleccionado() {
         int index = tblPublicaciones.convertRowIndexToModel(tblPublicaciones.getSelectedRow());
         if (index == -1) {
             Dialogs.mostrarMensajeError(this, "No ha seleccionado ningun elemento de la tabla!");
             return;
         }
-        Autor autorEditar = autores.get(index);
-        editarAutorForm.setAutorModificar(autorEditar);
-        FormUtils.cargarForm(editarAutorForm, this);
-    }
-
+        Publicacion publicacionEditar = publicaciones.get(index);
+        editarPublicacionForm.setPublicacionModificar(publicacionEditar);
+        FormUtils.cargarForm(editarPublicacionForm, this);
+    }    
+    
     private void initFormsConectados() {
-        this.editarAutorForm = new EditarAutorForm(this, this.clienteLoggeado);
+        this.editarPublicacionForm = new FrmEditarPublicacion(this, this.clienteLoggeado);
     }
 
     @Override
     public void actualizaFrame() {
-        cargarTablaAutores();
+        cargarTablaPublicaciones();
     }
 }
