+++ 03_Implementation/EditandoMundos/src/main/java/org/itson/presentacion/FrmEliminarAutores.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.logging.Logger;
 import javax.swing.JFrame;
 import javax.swing.table.DefaultTableModel;
-import org.itson.controladores.ControladorPublicacion;
-import org.itson.dominio.Publicacion;
+import org.itson.controladores.ControladorAutor;
+import org.itson.dominio.Autor;
 import org.itson.dominio.Usuario;
 import org.itson.utils.ConfiguracionPaginado;
 import org.itson.utils.Dialogs;
@@ -15,41 +15,39 @@ import org.itson.utils.FormUtils;
  *
  * @author Toled
  */
-public class EditarPublicacionesForm extends JFrameActualizable{
+public class FrmEliminarAutores extends JFrameActualizable {
 
-    private static final Logger LOG = Logger.getLogger(EditarPublicacionesForm.class.getName());
-    private Usuario clienteLoggeado;
+    private static final Logger LOG = Logger.getLogger(FrmEliminarAutores.class.getName());
     private ConfiguracionPaginado configPaginado;
+    private Usuario clienteLoggeado;
     private final JFrame frmAnterior;
-    private EditarPublicacionForm editarPublicacionForm;
-    private List<Publicacion> publicaciones;
+    private List<Autor> autores;
 
-    public EditarPublicacionesForm(JFrame frmAnterior, Usuario clienteLoggeado) {
+    public FrmEliminarAutores(JFrame frmAnterior, Usuario clienteLoggeado) {
         initComponents();
         this.configPaginado = new ConfiguracionPaginado(this.tblPublicaciones.getModel().getRowCount(), 0);
         this.clienteLoggeado = clienteLoggeado;
         this.frmAnterior = frmAnterior;
-        cargarTablaPublicaciones();
-        initFormsConectados();
+        cargarTablaAutores();
     }
 
-    public final void cargarTablaPublicaciones() {
+    public final void cargarTablaAutores() {
 
-        publicaciones = this.conseguirListaPublicaciones();
+        autores = this.conseguirListaAutores();
 
         DefaultTableModel modeloTabla = (DefaultTableModel) this.tblPublicaciones.getModel();
         modeloTabla.setRowCount(0);
-        for (Publicacion publicacion : publicaciones) {
+        String nombreCompleto;
+        for (Autor autor : autores) {
+            nombreCompleto = autor.getNombres() + " " + autor.getApellidoPaterno() + " " + autor.getApellidoMaterno();
             Object[] fila = {
-                publicacion.getTitulo(),
-                publicacion.getAutor().getNombres() + " " + publicacion.getAutor().getApellidoPaterno(),
-                publicacion.getNoPaginas(),
-                "$" + publicacion.getCostoProd(),
-                "$" + publicacion.getCostoVenta()};
+                nombreCompleto,
+                autor.getEdad(),
+                autor.getNacionalidad()
+            };
 
             modeloTabla.addRow(fila);
         }
-
     }
 
     @SuppressWarnings("unchecked")
@@ -63,8 +61,8 @@ public class EditarPublicacionesForm extends JFrameActualizable{
         panelTablaCuentas = new javax.swing.JScrollPane();
         tblPublicaciones = new javax.swing.JTable();
         btnAdelante = new javax.swing.JButton();
+        btnEliminar = new javax.swing.JButton();
         btnRetroceder = new javax.swing.JButton();
-        btnEditar = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setMinimumSize(new java.awt.Dimension(600, 400));
@@ -78,7 +76,7 @@ public class EditarPublicacionesForm extends JFrameActualizable{
 
         lblOperaciones.setFont(new java.awt.Font("Nirmala UI Semilight", 0, 24)); // NOI18N
         lblOperaciones.setForeground(new java.awt.Color(255, 255, 255));
-        lblOperaciones.setText("Editar Publicaciones");
+        lblOperaciones.setText("Eliminar Autores");
 
         btnAtras.setBackground(new java.awt.Color(0, 102, 255));
         btnAtras.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
@@ -99,9 +97,9 @@ public class EditarPublicacionesForm extends JFrameActualizable{
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
                 .addGap(20, 20, 20)
                 .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(139, 139, 139)
+                .addGap(128, 128, 128)
                 .addComponent(lblOperaciones)
-                .addContainerGap(174, Short.MAX_VALUE))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -117,26 +115,26 @@ public class EditarPublicacionesForm extends JFrameActualizable{
 
         tblPublicaciones.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null},
-                {null, null, null, null, null}
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null},
+                {null, null, null}
             },
             new String [] {
-                "Titulo", "Autor", "No. Paginas", "Producción", "Venta"
+                "Nombre", "Edad", "Nacionalidad"
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class
+                java.lang.String.class, java.lang.String.class, java.lang.String.class
             };
             boolean[] canEdit = new boolean [] {
-                false, false, false, false, true
+                false, false, false
             };
 
             public Class getColumnClass(int columnIndex) {
@@ -149,6 +147,11 @@ public class EditarPublicacionesForm extends JFrameActualizable{
         });
         panelTablaCuentas.setViewportView(tblPublicaciones);
         tblPublicaciones.getColumnModel().getSelectionModel().setSelectionMode(javax.swing.ListSelectionModel.SINGLE_INTERVAL_SELECTION);
+        if (tblPublicaciones.getColumnModel().getColumnCount() > 0) {
+            tblPublicaciones.getColumnModel().getColumn(0).setResizable(false);
+            tblPublicaciones.getColumnModel().getColumn(1).setResizable(false);
+            tblPublicaciones.getColumnModel().getColumn(2).setResizable(false);
+        }
 
         Background.add(panelTablaCuentas, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 150, 560, 190));
 
@@ -165,6 +168,19 @@ public class EditarPublicacionesForm extends JFrameActualizable{
         });
         Background.add(btnAdelante, new org.netbeans.lib.awtextra.AbsoluteConstraints(540, 340, 30, 30));
 
+        btnEliminar.setBackground(new java.awt.Color(0, 102, 255));
+        btnEliminar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
+        btnEliminar.setForeground(new java.awt.Color(255, 255, 255));
+        btnEliminar.setText("Eliminar");
+        btnEliminar.setBorder(null);
+        btnEliminar.setBorderPainted(false);
+        btnEliminar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnEliminarActionPerformed(evt);
+            }
+        });
+        Background.add(btnEliminar, new org.netbeans.lib.awtextra.AbsoluteConstraints(230, 360, 140, 30));
+
         btnRetroceder.setBackground(new java.awt.Color(0, 102, 255));
         btnRetroceder.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
         btnRetroceder.setForeground(new java.awt.Color(255, 255, 255));
@@ -177,19 +193,6 @@ public class EditarPublicacionesForm extends JFrameActualizable{
             }
         });
         Background.add(btnRetroceder, new org.netbeans.lib.awtextra.AbsoluteConstraints(510, 340, 30, 30));
-
-        btnEditar.setBackground(new java.awt.Color(0, 102, 255));
-        btnEditar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
-        btnEditar.setForeground(new java.awt.Color(255, 255, 255));
-        btnEditar.setText("Editar");
-        btnEditar.setBorder(null);
-        btnEditar.setBorderPainted(false);
-        btnEditar.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnEditarActionPerformed(evt);
-            }
-        });
-        Background.add(btnEditar, new org.netbeans.lib.awtextra.AbsoluteConstraints(250, 350, 150, 30));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -220,28 +223,29 @@ public class EditarPublicacionesForm extends JFrameActualizable{
      */
     private void btnAdelanteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAdelanteActionPerformed
         this.configPaginado.avanzarPag();
-        this.cargarTablaPublicaciones();
+        this.cargarTablaAutores();
     }//GEN-LAST:event_btnAdelanteActionPerformed
     /**
      * Retrocede en la pagina de operaciones
      *
      * @param evt Evento que lo acciono
      */
+    private void btnEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEliminarActionPerformed
+        this.eliminarElementoSeleccionado();
+        this.cargarTablaAutores();
+    }//GEN-LAST:event_btnEliminarActionPerformed
+
     private void btnRetrocederActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRetrocederActionPerformed
         this.configPaginado.retrocederPag();
-        this.cargarTablaPublicaciones();
+        this.cargarTablaAutores();
     }//GEN-LAST:event_btnRetrocederActionPerformed
-
-    private void btnEditarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEditarActionPerformed
-        this.editarElementoSeleccionado();
-    }//GEN-LAST:event_btnEditarActionPerformed
 
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JPanel Background;
     private javax.swing.JButton btnAdelante;
     private javax.swing.JButton btnAtras;
-    private javax.swing.JButton btnEditar;
+    private javax.swing.JButton btnEliminar;
     private javax.swing.JButton btnRetroceder;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JLabel lblOperaciones;
@@ -249,31 +253,30 @@ public class EditarPublicacionesForm extends JFrameActualizable{
     private javax.swing.JTable tblPublicaciones;
     // End of variables declaration//GEN-END:variables
 
-    private List<Publicacion> conseguirListaPublicaciones() {
-        return ControladorPublicacion.consultaPaginado(this.configPaginado);
+    private List<Autor> conseguirListaAutores() {
+        return ControladorAutor.consultaPaginado(this.configPaginado);
     }
 
     private void regresar() {
         FormUtils.regresar(frmAnterior, this);
     }
-    
-    private void editarElementoSeleccionado() {
+
+    private void eliminarElementoSeleccionado() {
         int index = tblPublicaciones.convertRowIndexToModel(tblPublicaciones.getSelectedRow());
         if (index == -1) {
             Dialogs.mostrarMensajeError(this, "No ha seleccionado ningun elemento de la tabla!");
             return;
         }
-        Publicacion publicacionEditar = publicaciones.get(index);
-        editarPublicacionForm.setPublicacionModificar(publicacionEditar);
-        FormUtils.cargarForm(editarPublicacionForm, this);
-    }    
-    
-    private void initFormsConectados() {
-        this.editarPublicacionForm = new EditarPublicacionForm(this, this.clienteLoggeado);
+
+        int eleccion = Dialogs.mostrarMensajeYesNoOption(this, "¿Seguro que desea eliminar el autor seleccionado?", "Confirmación");
+        if (eleccion == Dialogs.OPCION_SI) {
+            Autor autorEliminar = autores.get(index);
+            ControladorAutor.eliminarAutor(autorEliminar);
+        }
     }
 
     @Override
     public void actualizaFrame() {
-        cargarTablaPublicaciones();
+        cargarTablaAutores();
     }
 }
