+++ 03_Implementation/EditandoMundos/src/main/java/org/itson/presentacion/FrmEliminarPublicaciones.java
@@ -4,49 +4,46 @@ import java.util.List;
 import java.util.logging.Logger;
 import javax.swing.JFrame;
 import javax.swing.table.DefaultTableModel;
-import org.itson.controladores.ControladorAutor;
-import org.itson.dominio.Autor;
+import org.itson.controladores.ControladorPublicacion;
+import org.itson.dominio.Publicacion;
 import org.itson.dominio.Usuario;
 import org.itson.utils.ConfiguracionPaginado;
+import org.itson.utils.Dialogs;
 import org.itson.utils.FormUtils;
 
 /**
  *
  * @author Toled
  */
-public class AutoresForm extends JFrameActualizable {
+public class FrmEliminarPublicaciones extends JFrameActualizable{
 
-    private static final Logger LOG = Logger.getLogger(AutoresForm.class.getName());
-    private ConfiguracionPaginado configPaginado;
+    private static final Logger LOG = Logger.getLogger(FrmEliminarPublicaciones.class.getName());
     private Usuario clienteLoggeado;
+    private ConfiguracionPaginado configPaginado;
     private final JFrame frmAnterior;
+    private List<Publicacion> publicaciones;
 
-    public AutoresForm(JFrame frmAnterior, Usuario clienteLoggeado) {
+    public FrmEliminarPublicaciones(JFrame frmAnterior, Usuario clienteLoggeado) {
         initComponents();
         this.configPaginado = new ConfiguracionPaginado(this.tblPublicaciones.getModel().getRowCount(), 0);
         this.clienteLoggeado = clienteLoggeado;
         this.frmAnterior = frmAnterior;
-        cargarTablaAutores();
+        cargarTablaPublicaciones();
     }
 
-    public final void cargarTablaAutores() {
+    public final void cargarTablaPublicaciones() {
 
-        List<Autor> listaAutores = this.conseguirListaAutores();
+        publicaciones = this.conseguirListaPublicaciones();
 
         DefaultTableModel modeloTabla = (DefaultTableModel) this.tblPublicaciones.getModel();
         modeloTabla.setRowCount(0);
-        for (Autor autor : listaAutores) {
-            // TODO utils generar nombre completo
-            // TODO apellido materno puede no existir.
-            String nombreCompleto
-                    = autor.getNombres() + " "
-                    + autor.getApellidoPaterno()
-                    + " " + autor.getApellidoMaterno();
+        for (Publicacion publicacion : publicaciones) {
             Object[] fila = {
-                nombreCompleto,
-                autor.getEdad(),
-                autor.getNacionalidad()
-            };
+                publicacion.getTitulo(),
+                publicacion.getAutor().getNombres() + " " + publicacion.getAutor().getApellidoPaterno(),
+                publicacion.getNoPaginas(),
+                "$" + publicacion.getCostoProd(),
+                "$" + publicacion.getCostoVenta()};
 
             modeloTabla.addRow(fila);
         }
@@ -65,6 +62,7 @@ public class AutoresForm extends JFrameActualizable {
         tblPublicaciones = new javax.swing.JTable();
         btnAdelante = new javax.swing.JButton();
         btnRetroceder = new javax.swing.JButton();
+        btnEliminar = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setMinimumSize(new java.awt.Dimension(600, 400));
@@ -78,7 +76,7 @@ public class AutoresForm extends JFrameActualizable {
 
         lblOperaciones.setFont(new java.awt.Font("Nirmala UI Semilight", 0, 24)); // NOI18N
         lblOperaciones.setForeground(new java.awt.Color(255, 255, 255));
-        lblOperaciones.setText("Consultar Autores");
+        lblOperaciones.setText("Eliminar Publicaciones");
 
         btnAtras.setBackground(new java.awt.Color(0, 102, 255));
         btnAtras.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
@@ -101,7 +99,7 @@ public class AutoresForm extends JFrameActualizable {
                 .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(128, 128, 128)
                 .addComponent(lblOperaciones)
-                .addContainerGap(204, Short.MAX_VALUE))
+                .addContainerGap(164, Short.MAX_VALUE))
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -110,33 +108,33 @@ public class AutoresForm extends JFrameActualizable {
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblOperaciones)
                     .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(22, Short.MAX_VALUE))
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
         );
 
         Background.add(jPanel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 600, 90));
 
         tblPublicaciones.setModel(new javax.swing.table.DefaultTableModel(
             new Object [][] {
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null},
-                {null, null, null}
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null},
+                {null, null, null, null, null}
             },
             new String [] {
-                "Nombre", "Edad", "Nacionalidad"
+                "Titulo", "Autor", "No. Paginas", "Producción", "Venta"
             }
         ) {
             Class[] types = new Class [] {
-                java.lang.String.class, java.lang.String.class, java.lang.String.class
+                java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class, java.lang.String.class
             };
             boolean[] canEdit = new boolean [] {
-                false, false, false
+                false, false, false, false, true
             };
 
             public Class getColumnClass(int columnIndex) {
@@ -147,17 +145,8 @@ public class AutoresForm extends JFrameActualizable {
                 return canEdit [columnIndex];
             }
         });
-        tblPublicaciones.setColumnSelectionAllowed(true);
         panelTablaCuentas.setViewportView(tblPublicaciones);
         tblPublicaciones.getColumnModel().getSelectionModel().setSelectionMode(javax.swing.ListSelectionModel.SINGLE_INTERVAL_SELECTION);
-        if (tblPublicaciones.getColumnModel().getColumnCount() > 0) {
-            tblPublicaciones.getColumnModel().getColumn(0).setResizable(false);
-            tblPublicaciones.getColumnModel().getColumn(0).setHeaderValue("Nombre");
-            tblPublicaciones.getColumnModel().getColumn(1).setResizable(false);
-            tblPublicaciones.getColumnModel().getColumn(1).setHeaderValue("Edad");
-            tblPublicaciones.getColumnModel().getColumn(2).setResizable(false);
-            tblPublicaciones.getColumnModel().getColumn(2).setHeaderValue("Nacionalidad");
-        }
 
         Background.add(panelTablaCuentas, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 150, 560, 190));
 
@@ -186,6 +175,19 @@ public class AutoresForm extends JFrameActualizable {
             }
         });
         Background.add(btnRetroceder, new org.netbeans.lib.awtextra.AbsoluteConstraints(510, 340, 30, 30));
+
+        btnEliminar.setBackground(new java.awt.Color(0, 102, 255));
+        btnEliminar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
+        btnEliminar.setForeground(new java.awt.Color(255, 255, 255));
+        btnEliminar.setText("Eliminar");
+        btnEliminar.setBorder(null);
+        btnEliminar.setBorderPainted(false);
+        btnEliminar.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnEliminarActionPerformed(evt);
+            }
+        });
+        Background.add(btnEliminar, new org.netbeans.lib.awtextra.AbsoluteConstraints(250, 350, 150, 30));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -216,7 +218,7 @@ public class AutoresForm extends JFrameActualizable {
      */
     private void btnAdelanteActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnAdelanteActionPerformed
         this.configPaginado.avanzarPag();
-        this.cargarTablaAutores();
+        this.cargarTablaPublicaciones();
     }//GEN-LAST:event_btnAdelanteActionPerformed
     /**
      * Retrocede en la pagina de operaciones
@@ -225,13 +227,20 @@ public class AutoresForm extends JFrameActualizable {
      */
     private void btnRetrocederActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRetrocederActionPerformed
         this.configPaginado.retrocederPag();
-        this.cargarTablaAutores();
+        this.cargarTablaPublicaciones();
     }//GEN-LAST:event_btnRetrocederActionPerformed
+
+    private void btnEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEliminarActionPerformed
+        this.eliminarElementoSeleccionado();
+        this.cargarTablaPublicaciones();
+    }//GEN-LAST:event_btnEliminarActionPerformed
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JPanel Background;
     private javax.swing.JButton btnAdelante;
     private javax.swing.JButton btnAtras;
+    private javax.swing.JButton btnEliminar;
     private javax.swing.JButton btnRetroceder;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JLabel lblOperaciones;
@@ -239,16 +248,30 @@ public class AutoresForm extends JFrameActualizable {
     private javax.swing.JTable tblPublicaciones;
     // End of variables declaration//GEN-END:variables
 
-    private List<Autor> conseguirListaAutores() {
-        return ControladorAutor.consultaPaginado(this.configPaginado);
+    private List<Publicacion> conseguirListaPublicaciones() {
+        return ControladorPublicacion.consultaPaginado(this.configPaginado);
     }
 
     private void regresar() {
         FormUtils.regresar(frmAnterior, this);
     }
+    
+    private void eliminarElementoSeleccionado() {
+        int index = tblPublicaciones.convertRowIndexToModel(tblPublicaciones.getSelectedRow());
+        if (index == -1) {
+            Dialogs.mostrarMensajeError(this, "No ha seleccionado ningun elemento de la tabla!");
+            return;
+        }
+
+        int eleccion = Dialogs.mostrarMensajeYesNoOption(this, "¿Seguro que desea eliminar la publicación seleccionada?", "Confirmación");
+        if (eleccion == Dialogs.OPCION_SI) {
+            Publicacion publicacionEliminar = publicaciones.get(index);
+            ControladorPublicacion.eliminarPublicacion(publicacionEliminar);
+        }
+    }    
 
     @Override
     public void actualizaFrame() {
-        cargarTablaAutores();
+        cargarTablaPublicaciones();
     }
 }
