+++ 03_Implementation/EditandoMundos/src/main/java/org/itson/presentacion/FrmIniciarSesion.java
@@ -6,18 +6,19 @@ import org.itson.repositories.UsuariosRepository;
 import org.itson.utils.Dialogs;
 import static org.itson.utils.Dialogs.mostrarMensajeError;
 import org.itson.utils.Encriptador;
+import org.itson.utils.FormUtils;
 import static org.itson.utils.ValidacionesForms.isValidText;
 
 /**
  *
  * @author Toled
  */
-public class IniciarSesionForm extends javax.swing.JFrame {
+public class FrmIniciarSesion extends javax.swing.JFrame {
 
     private UnitOfWork unitOfWork;
     private Usuario usuarioLoggeado;
 
-    public IniciarSesionForm() {
+    public FrmIniciarSesion() {
         initComponents();
         unitOfWork = new UnitOfWork();
     }
@@ -36,8 +37,9 @@ public class IniciarSesionForm extends javax.swing.JFrame {
         jSeparator2 = new javax.swing.JSeparator();
         lblContraseña = new javax.swing.JLabel();
         campoTextoContraseña = new javax.swing.JPasswordField();
-        btnIngresar = new javax.swing.JButton();
+        btnIniciarSesion = new javax.swing.JButton();
         txtIniciarSesion = new javax.swing.JLabel();
+        btnRegistrarse = new javax.swing.JButton();
 
         jButton2.setText("jButton2");
 
@@ -97,23 +99,37 @@ public class IniciarSesionForm extends javax.swing.JFrame {
         campoTextoContraseña.setBorder(null);
         Background.add(campoTextoContraseña, new org.netbeans.lib.awtextra.AbsoluteConstraints(210, 260, 190, 20));
 
-        btnIngresar.setBackground(new java.awt.Color(0, 102, 204));
-        btnIngresar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 12)); // NOI18N
-        btnIngresar.setForeground(new java.awt.Color(255, 255, 255));
-        btnIngresar.setText("Ingresar");
-        btnIngresar.setBorder(null);
-        btnIngresar.setBorderPainted(false);
-        btnIngresar.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
-        btnIngresar.addActionListener(new java.awt.event.ActionListener() {
+        btnIniciarSesion.setBackground(new java.awt.Color(0, 102, 204));
+        btnIniciarSesion.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 12)); // NOI18N
+        btnIniciarSesion.setForeground(new java.awt.Color(255, 255, 255));
+        btnIniciarSesion.setText("Iniciar Sesión");
+        btnIniciarSesion.setBorder(null);
+        btnIniciarSesion.setBorderPainted(false);
+        btnIniciarSesion.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        btnIniciarSesion.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnIngresarActionPerformed(evt);
+                btnIniciarSesionActionPerformed(evt);
             }
         });
-        Background.add(btnIngresar, new org.netbeans.lib.awtextra.AbsoluteConstraints(270, 310, 70, 30));
+        Background.add(btnIniciarSesion, new org.netbeans.lib.awtextra.AbsoluteConstraints(310, 300, 90, 30));
 
         txtIniciarSesion.setFont(new java.awt.Font("Nirmala UI Semilight", 0, 16)); // NOI18N
         txtIniciarSesion.setText("Iniciar sesion");
         Background.add(txtIniciarSesion, new org.netbeans.lib.awtextra.AbsoluteConstraints(260, 110, -1, -1));
+
+        btnRegistrarse.setBackground(new java.awt.Color(0, 102, 204));
+        btnRegistrarse.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 12)); // NOI18N
+        btnRegistrarse.setForeground(new java.awt.Color(255, 255, 255));
+        btnRegistrarse.setText("Registrarse");
+        btnRegistrarse.setBorder(null);
+        btnRegistrarse.setBorderPainted(false);
+        btnRegistrarse.setCursor(new java.awt.Cursor(java.awt.Cursor.DEFAULT_CURSOR));
+        btnRegistrarse.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                btnRegistrarseActionPerformed(evt);
+            }
+        });
+        Background.add(btnRegistrarse, new org.netbeans.lib.awtextra.AbsoluteConstraints(200, 300, 90, 30));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -135,18 +151,23 @@ public class IniciarSesionForm extends javax.swing.JFrame {
      *
      * @param evt Evento que lo acciona
      */
-    private void btnIngresarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnIngresarActionPerformed
+    private void btnIniciarSesionActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnIniciarSesionActionPerformed
         this.iniciarSesion();
-    }//GEN-LAST:event_btnIngresarActionPerformed
+    }//GEN-LAST:event_btnIniciarSesionActionPerformed
 
     private void campoTextoUsuarioActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_campoTextoUsuarioActionPerformed
         // TODO add your handling code here:
     }//GEN-LAST:event_campoTextoUsuarioActionPerformed
 
+    private void btnRegistrarseActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnRegistrarseActionPerformed
+        this.registrarse();
+    }//GEN-LAST:event_btnRegistrarseActionPerformed
+
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JPanel Background;
-    private javax.swing.JButton btnIngresar;
+    private javax.swing.JButton btnIniciarSesion;
+    private javax.swing.JButton btnRegistrarse;
     private javax.swing.JPasswordField campoTextoContraseña;
     private javax.swing.JTextField campoTextoUsuario;
     private javax.swing.JButton jButton2;
@@ -169,12 +190,6 @@ public class IniciarSesionForm extends javax.swing.JFrame {
         } catch (NoResultException e) {
             mostrarMensajeError(rootPane, "No se encontró el usuario");
         }
-    }
-
-    private void cargarMenuPrincipal() {
-        MenuPrincipalForm menuPrincipal = new MenuPrincipalForm(usuarioLoggeado);
-        menuPrincipal.setVisible(true);
-        this.dispose();
     }
 
     private boolean validarCampos() {
@@ -201,5 +216,20 @@ public class IniciarSesionForm extends javax.swing.JFrame {
         this.usuarioLoggeado = usuario;
 
         this.cargarMenuPrincipal();
+    }
+
+    private void cargarMenuPrincipal() {
+        MenuPrincipalForm menuPrincipal = new MenuPrincipalForm(usuarioLoggeado);
+        FormUtils.cargarForm(menuPrincipal, this);
+    }
+
+    private void registrarse() {
+        this.cargarRegistroCliente();
+    }
+
+    private void cargarRegistroCliente() {
+//        FrmRegistroCliente frmRegistroCliente = new FrmRegistroCliente(usuarioLoggeado);
+//        frmRegistroCliente.setVisible(true);
+//        this.dispose();
     }
 }
