+++ 03_Implementation/EditandoMundos/src/main/java/org/itson/presentacion/FrmPublicaciones.java
@@ -8,22 +8,20 @@ import org.itson.controladores.ControladorPublicacion;
 import org.itson.dominio.Publicacion;
 import org.itson.dominio.Usuario;
 import org.itson.utils.ConfiguracionPaginado;
-import org.itson.utils.Dialogs;
 import org.itson.utils.FormUtils;
 
 /**
  *
  * @author Toled
  */
-public class EliminarPublicacionesForm extends JFrameActualizable{
+public class FrmPublicaciones extends JFrame {
 
-    private static final Logger LOG = Logger.getLogger(EliminarPublicacionesForm.class.getName());
+    private static final Logger LOG = Logger.getLogger(FrmPublicaciones.class.getName());
     private Usuario clienteLoggeado;
     private ConfiguracionPaginado configPaginado;
     private final JFrame frmAnterior;
-    private List<Publicacion> publicaciones;
 
-    public EliminarPublicacionesForm(JFrame frmAnterior, Usuario clienteLoggeado) {
+    public FrmPublicaciones(JFrame frmAnterior, Usuario clienteLoggeado) {
         initComponents();
         this.configPaginado = new ConfiguracionPaginado(this.tblPublicaciones.getModel().getRowCount(), 0);
         this.clienteLoggeado = clienteLoggeado;
@@ -33,7 +31,7 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
 
     public final void cargarTablaPublicaciones() {
 
-        publicaciones = this.conseguirListaPublicaciones();
+        List<Publicacion> publicaciones = this.conseguirListaPublicaciones();
 
         DefaultTableModel modeloTabla = (DefaultTableModel) this.tblPublicaciones.getModel();
         modeloTabla.setRowCount(0);
@@ -62,7 +60,6 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
         tblPublicaciones = new javax.swing.JTable();
         btnAdelante = new javax.swing.JButton();
         btnRetroceder = new javax.swing.JButton();
-        btnEliminar = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
         setMinimumSize(new java.awt.Dimension(600, 400));
@@ -76,7 +73,7 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
 
         lblOperaciones.setFont(new java.awt.Font("Nirmala UI Semilight", 0, 24)); // NOI18N
         lblOperaciones.setForeground(new java.awt.Color(255, 255, 255));
-        lblOperaciones.setText("Eliminar Publicaciones");
+        lblOperaciones.setText("Publicaciones");
 
         btnAtras.setBackground(new java.awt.Color(0, 102, 255));
         btnAtras.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
@@ -97,9 +94,9 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel1Layout.createSequentialGroup()
                 .addGap(20, 20, 20)
                 .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(128, 128, 128)
+                .addGap(168, 168, 168)
                 .addComponent(lblOperaciones)
-                .addContainerGap(164, Short.MAX_VALUE))
+                .addContainerGap(211, Short.MAX_VALUE))
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -108,7 +105,7 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
                 .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
                     .addComponent(lblOperaciones)
                     .addComponent(btnAtras, javax.swing.GroupLayout.PREFERRED_SIZE, 40, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+                .addContainerGap(22, Short.MAX_VALUE))
         );
 
         Background.add(jPanel1, new org.netbeans.lib.awtextra.AbsoluteConstraints(0, 0, 600, 90));
@@ -145,8 +142,7 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
                 return canEdit [columnIndex];
             }
         });
-        tblPublicaciones.setCellSelectionEnabled(false);
-        tblPublicaciones.setRowSelectionAllowed(true);
+        tblPublicaciones.setColumnSelectionAllowed(true);
         panelTablaCuentas.setViewportView(tblPublicaciones);
         tblPublicaciones.getColumnModel().getSelectionModel().setSelectionMode(javax.swing.ListSelectionModel.SINGLE_INTERVAL_SELECTION);
 
@@ -177,19 +173,6 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
             }
         });
         Background.add(btnRetroceder, new org.netbeans.lib.awtextra.AbsoluteConstraints(510, 340, 30, 30));
-
-        btnEliminar.setBackground(new java.awt.Color(0, 102, 255));
-        btnEliminar.setFont(new java.awt.Font("Nirmala UI Semilight", 1, 14)); // NOI18N
-        btnEliminar.setForeground(new java.awt.Color(255, 255, 255));
-        btnEliminar.setText("Eliminar");
-        btnEliminar.setBorder(null);
-        btnEliminar.setBorderPainted(false);
-        btnEliminar.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                btnEliminarActionPerformed(evt);
-            }
-        });
-        Background.add(btnEliminar, new org.netbeans.lib.awtextra.AbsoluteConstraints(250, 350, 150, 30));
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
         getContentPane().setLayout(layout);
@@ -232,17 +215,11 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
         this.cargarTablaPublicaciones();
     }//GEN-LAST:event_btnRetrocederActionPerformed
 
-    private void btnEliminarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_btnEliminarActionPerformed
-        this.eliminarElementoSeleccionado();
-        this.cargarTablaPublicaciones();
-    }//GEN-LAST:event_btnEliminarActionPerformed
-
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JPanel Background;
     private javax.swing.JButton btnAdelante;
     private javax.swing.JButton btnAtras;
-    private javax.swing.JButton btnEliminar;
     private javax.swing.JButton btnRetroceder;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JLabel lblOperaciones;
@@ -256,24 +233,5 @@ public class EliminarPublicacionesForm extends JFrameActualizable{
 
     private void regresar() {
         FormUtils.regresar(frmAnterior, this);
-    }
-    
-    private void eliminarElementoSeleccionado() {
-        int index = tblPublicaciones.convertRowIndexToModel(tblPublicaciones.getSelectedRow());
-        if (index == -1) {
-            Dialogs.mostrarMensajeError(this, "No ha seleccionado ningun elemento de la tabla!");
-            return;
-        }
-
-        int eleccion = Dialogs.mostrarMensajeYesNoOption(this, "¿Seguro que desea eliminar la publicación seleccionada?", "Confirmación");
-        if (eleccion == Dialogs.OPCION_SI) {
-            Publicacion publicacionEliminar = publicaciones.get(index);
-            ControladorPublicacion.eliminarPublicacion(publicacionEliminar);
-        }
-    }    
-
-    @Override
-    public void actualizaFrame() {
-        cargarTablaPublicaciones();
     }
 }
